+++ src/hooks/mpi/SharedLibrary.hpp
@@ -8,8 +8,8 @@
  *
  */
 
-#ifndef sarus_hooks_mpi_Utils_hpp
-#define sarus_hooks_mpi_Utils_hpp
+#ifndef sarus_hooks_mpi_SharedLibrary_hpp
+#define sarus_hooks_mpi_SharedLibrary_hpp
 
 #include <boost/filesystem.hpp>
 #include <vector>
@@ -23,26 +23,33 @@ namespace hooks {
 namespace mpi {
 
 class SharedLibrary {
+
     // Using naming convention mentioned in The Linux Programming Interface book.
 public:
     SharedLibrary(const boost::filesystem::path& path, const boost::filesystem::path& rootDir="");
 
     bool hasMajorVersion() const;
-    bool isFullAbiCompatible(const SharedLibrary& sl) const;
-    bool isMajorAbiCompatible(const SharedLibrary& sl) const;
+    bool isFullAbiCompatible(const SharedLibrary& hostLibrary) const;
+    bool isMajorAbiCompatible(const SharedLibrary& hostLibrary) const;
     SharedLibrary pickNewestAbiCompatibleLibrary(const std::vector<SharedLibrary>& candidates) const;
     std::string getLinkerName() const;
     boost::filesystem::path getPath() const;
     std::string getRealName() const;
+    const boost::optional<int>& getMajorVersion() const { return major; }
+    const boost::optional<int>& getMinorVersion() const { return minor; }
 
 private:
-    int major = 0;
-    int minor = 0;
-    int patch = 0;
+    boost::optional<int> major;
+    boost::optional<int> minor;
+    boost::optional<int> patch;
     std::string linkerName;
     boost::filesystem::path path;
     std::string realName;
 };
+
+bool areMajorAbiCompatible(const SharedLibrary&, const SharedLibrary&);
+bool areFullAbiCompatible(const SharedLibrary&, const SharedLibrary&);
+bool areStrictlyAbiCompatible(const SharedLibrary&, const SharedLibrary&);
 
 }}}